20210109_型の変更(short→byte)

+++ Assets/Scripts/GenerateManager.cs
@@ -7,52 +7,52 @@ public class GenerateManager : MonoBehaviour
     // ブロックのプレハブ
     public GameObject brockPrefab;
 
-    short[][] blocks =
+    byte[][] blocks =
         {
-            new short [] { 0, 1, 1, 1, 1, 1, 1, 1, 1, 1 },
-            new short [] { 0, 0, 0, 0, 0, 0, 0, 1, 1, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 1, 1, 0, 0 },
-            new short [] { 0, 0, 0, 0, 0, 1, 1, 0, 0, 0 },
-            new short [] { 0, 0, 0, 0, 1, 1, 0, 0, 0, 0 },
-            new short [] { 0, 0, 0, 1, 1, 0, 0, 0, 0, 0 },
-            new short [] { 0, 1, 1, 1, 0, 0, 0, 0, 0, 0 },
-            new short [] { 1, 1, 1, 1, 1, 1, 1, 1, 1, 0 },
-            new short [] { 0, 0, 0, 1, 1, 0, 0, 1, 0, 0 },
-            new short [] { 0, 0, 0, 0, 1, 0, 0, 1, 0, 0 },
-            new short [] { 0, 0, 0, 0, 1, 0, 0, 1, 0, 0 },
-            new short [] { 0, 0, 0, 0, 1, 1, 1, 1, 0, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 1, 1, 0, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 1, 1, 0, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 1, 1, 0, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 1, 1, 0, 0 },
+            new byte [] { 0, 1, 1, 1, 1, 1, 1, 1, 1, 1 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 0, 1, 1, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 1, 1, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 1, 1, 0, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 1, 1, 0, 0, 0, 0 },
+            new byte [] { 0, 0, 0, 1, 1, 0, 0, 0, 0, 0 },
+            new byte [] { 0, 1, 1, 1, 0, 0, 0, 0, 0, 0 },
+            new byte [] { 1, 1, 1, 1, 1, 1, 1, 1, 1, 0 },
+            new byte [] { 0, 0, 0, 1, 1, 0, 0, 1, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 1, 0, 0, 1, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 1, 0, 0, 1, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 1, 1, 1, 1, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 1, 1, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 1, 1, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 1, 1, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 1, 1, 0, 0 },
 
         };
 
-    short[][] blocks2 =
+    byte[][] blocks2 =
         {
-            new short [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
-            new short [] { 0, 1, 1, 0, 0, 0, 0, 0, 0, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 0, 0, 1, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
-            new short [] { 0, 0, 0, 1, 1, 0, 0, 0, 0, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
-            new short [] { 0, 0, 1, 1, 0, 0, 0, 0, 0, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 0, 1, 0, 0 },
-            new short [] { 0, 0, 0, 0, 1, 0, 0, 1, 0, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 1, 0, 0, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
-            new short [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
+            new byte [] { 0, 1, 1, 0, 0, 0, 0, 0, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 0, 0, 1, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
+            new byte [] { 0, 0, 0, 1, 1, 0, 0, 0, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
+            new byte [] { 0, 0, 1, 1, 0, 0, 0, 0, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 0, 1, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 1, 0, 0, 1, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 1, 0, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
+            new byte [] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
 
         };
 
 
     // ブロックを生成する位置
-    short pos1 = 0;
-    short pos2 = 0;
+    int pos1 = 0;
+    int pos2 = 0;
 
 
     // Start is called before the first frame update
@@ -73,9 +73,9 @@ public class GenerateManager : MonoBehaviour
 
         if (pos1 < blocks.Length)
         {
-            for (short i = pos1; i < pos1 + 1; i++)
+            for (int i = pos1; i < pos1 + 1; i++)
             {
-                for (short j = 0; j < blocks[i].Length; j++)
+                for (int j = 0; j < blocks[i].Length; j++)
                 {
                     if (blocks[i][j] == 1)
                     {
@@ -96,9 +96,9 @@ public class GenerateManager : MonoBehaviour
 
         if (pos2 < blocks2.Length)
         {
-            for (short i = pos2; i < pos2 + 1; i++)
+            for (int i = pos2; i < pos2 + 1; i++)
             {
-                for (short j = 0; j < blocks2[i].Length; j++)
+                for (int j = 0; j < blocks2[i].Length; j++)
                 {
                     if (blocks2[i][j] == 1)
                     {
